+++ nuxt.config.ts
@@ -1,3 +1,5 @@
+import tailwindcss from "@tailwindcss/vite";
+
 // https://nuxt.com/docs/api/configuration/nuxt-config
 export default defineNuxtConfig({
   compatibilityDate: "2025-07-15",
@@ -5,4 +7,8 @@ export default defineNuxtConfig({
   typescript: {
     strict: true,
   },
+  vite: {
+    plugins: [tailwindcss() as any],
+  },
+  css: ["~/assets/css/main.css"],
 });